fix: update missionId type to Integer

+++ src/main/java/com/example/domain/MissionService.java
@@ -25,7 +25,7 @@ public class MissionService {
                 .toList();
     }
 
-    public Optional<MoonMission> getMissionById(Short id) {
+    public Optional<MoonMission> getMissionById(Integer id) {
         return missionRepository.findById(id);
     }
